move bg image to step2 remove marketing goals
